serdes/add_base_control: add auto_enable parameter to allow configuring startup behaviour.
On some systems where clocks are available at startup, it's convenient to automatically
enable the links. That's the default behaviour and what is used in LiteICLink bench. On
other systems where clocks are generated by an external chip that need to be configured,
we need to keep the SerDes in reset until the clock has been configured.