test_prbs: Handle PRBS error counter wrap around
Previously if the PRBS counter reached its max value of 0xffffffff,
the reported number of errors would drop to 0. And this happens
pretty quickly given the fast data rates.

Now we enable the new 'wrap' option of PRBSRX core so that the counter
wraps to 0 on overflows and we detect and handle that condition in
test_prbs.py

Signed-off-by: Sylvain Munaut <tnt@246tNt.com>